refactor the project
Refactoring the project is to use the same Cargo.toml, which is
convenient to manage dependencies.
