+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sealfs-rust"
+name = "sealfs"
 version = "0.1.0"
 edition = "2021"
 authors = ["The Sealfs Developers"]
@@ -19,16 +19,16 @@ tonic = "0.8.2"
 prost = "0.11.0"
 serde = { version = "1", features = ["derive"] }
 serde_yaml = "0.9.14"
-tonic-health = "0.7.1"
+# tonic-health = "0.7.1"
 dashmap = "5.4.0"
 async-trait = "0.1.59"
 nix = "0.26.1"
-
-client = {path = "client"}
-server = {path = "server"}
-manager = {path = "manager"}
-common = {path = "common"}
-rpc = {path = "rpc"}
+rocksdb = "0.19.0"
+bincode = "1.3.3"
+fasthash = "0.4.0"
+parking_lot = "0.12.1"
+fuser = "0.11.1"
+libc = "0.2"
 
 [build-dependencies]
 tonic-build = "0.8"
@@ -44,12 +44,3 @@ path = "src/bin/server.rs"
 [[bin]]
 name = "manager"
 path = "src/bin/manager.rs"
-
-[workspace]
-members = [
-    "server",
-    "client",
-    "common",
-    "manager",
-    "rpc",
-]
