+++ build.rs
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    tonic_build::compile_protos("proto/fsbase.proto")?;
     tonic_build::compile_protos("proto/manager.proto")?;
     Ok(())
 }
