--- client/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "client"
-version = "0.1.0"
-edition = "2021"
-authors = ["The Sealfs Developers"]
-license = "Apache-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio = { version = "1.21.2", features = ["full"] }
-libc = "0.2"
-fuser = "0.11.1"
-env_logger = "0.9"
-log = "0.4.17"
-clap = { version = "4.0.18", features = ["derive"] }
-lazy_static = "1.4.0"
-dashmap = "5.4.0"
-tonic = "0.8.2"
-serde = { version = "1", features = ["derive"] }
-serde_yaml = "0.9.14"
-fasthash = "0.4.0"
-parking_lot = "0.12.1"
-
-common = {path = "../common"}
-manager = {path = "../manager"}
-
-[dev-dependencies]
-env_logger = "0.9"
-log = "0.4.17"
-clap = { version = "4.0.18", features = ["derive"] }
-lazy_static = "1.4.0"
-dashmap = "5.4.0"
