--- common/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "common"
-version = "0.1.0"
-edition = "2021"
-authors = ["The Sealfs Developers"]
-license = "Apache-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tonic = "0.8.2"
-log = "0.4.17"
-lazy_static = "1.4.0"
-dashmap = "5.4.0"
-fasthash = "0.4.0"
-parking_lot = "0.12.1"
-
-manager = {path = "../manager"}
-
-
-
