+++ examples/manager.yaml
