--- manager/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "manager"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tonic = "0.8.2"
-log = "0.4.17"
-prost = "0.11.0"
-dashmap = "5.4.0"
-lazy_static = "1.4.0"
-tokio = { version = "1.21.2", features = ["full"] }
-
-[build-dependencies]
-tonic-build = "0.8"
