--- server/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "server"
-version = "0.1.0"
-edition = "2021"
-authors = ["The Sealfs Developers"]
-license = "Apache-2.0"
-
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio = { version = "1.21.2", features = ["full"] }
-tonic = "0.8.2"
-prost = "0.11.0"
-libc = "0.2"
-anyhow = "1.0.66"
-thiserror = "1.0.37"
-log = "0.4.17"
-rocksdb = { version = "0.19.0" }
-bincode = "1.3.3"
-serde = { version = "1.0.147", features = ["derive"] }
-dashmap = "5.4.0"
-nix = "0.26.1"
-lazy_static = "1.4.0"
-
-common = { path = "../common" }
-rpc = { path = "../rpc" }
-
-[build-dependencies]
-tonic-build = "0.8"
