+++ src/common/byte.rs
