+++ src/lib.rs
@@ -1,3 +1,9 @@
 // Copyright 2022 labring. All rights reserved.
 //
 // SPDX-License-Identifier: Apache-2.0
+
+pub mod client;
+pub mod common;
+pub mod manager;
+pub mod rpc;
+pub mod server;
